u2-d11 text analysis: fixed issues with Genius.com lyrics retrieval + rendered content.
TL;DR Genius.com changed their layout and it breaks existing (scraping) approaches for their retrieval. The R genius package that was used is now no longer maintained and will soon be removed from CRAN. I explored the use of another R package called geniusr, which uses the genius.com API, but it is currently broken as well due to the layout changes, and a patched version of it resulted in errors when retrieving the lyrics for some of the songs and not other. Long story short, I found an existing dataset of Ed Sheeran's lyrics, which is now downloaded and used by the code. I think it is better to eliminate the dependency on an external API for this purpose and focus more on the primary task of this notebook, which is to analyze text. The new data produces similarly-looking results.